+++ src/App.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import Kernable from 'components/kernable/kernable';
+
+import "styles.scss";
+
+function App() {
+  return (
+    <div className="App">
+      <Kernable word="Wave"/>
+      <Kernable word="Ashley"/>
+    </div>
+  );
+}
+
+export default App;