+++ src/index.js
@@ -1,43 +1,17 @@
-'use strict';
+import React from 'react';
+import ReactDOM from 'react-dom';
+import './index.css';
+import App from './App';
+import * as serviceWorker from './serviceWorker';
 
-import "./styles.css";
-import Painterro from 'painterro';
+ReactDOM.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>,
+  document.getElementById('root')
+);
 
-Painterro({
-	activeColor: '#00ff00', // default brush color is green
-	toolbarPosition: 'top',
-	hiddenTools: [
-		'select',
-		'settings',
-		'undo',
-		'redo',
-		'pixelize',
-		'crop',
-		//'eraser',
-		//'line',
-		'arrow',
-		'rect',
-		'ellipse',
-		'brush',
-		'text',
-		'rotate',
-		'resize',
-		'save',
-		'open',
-		'close'
-	],
-	colorScheme: {
-		main: '#414141', // make panels light-yellow
-		control: '#414141', // change controls color
-		controlContent: '#d9d9d9' // change controls color
-	}
-}).show();
-
-const UI = {
-  render: function(content, where = "beforeend") {
-	 const container = document.querySelector("#app");
-	 container.insertAdjacentHTML(where, content);
-  }
-};
-
-UI.render( "<h1>Learn Design</h1>", "afterbegin" );
+// If you want your app to work offline and load faster, you can change
+// unregister() to register() below. Note this comes with some pitfalls.
+// Learn more about service workers: https://bit.ly/CRA-PWA
+serviceWorker.unregister();