+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+- Adds new kernable component via #19
+- Switch to React and Sass via #19
+
 # 0.0.1
 
 - Initial setup of repo